requirements.txt
 google-api-python-client==2.129.0
 google-auth-httplib2==0.2.0
 google-auth-oauthlib==1.2.0
 black==24.3.0